feat(roles): implement roles module with CRUD operations and relationships

+++ ss1-backend-node/src/core/database/scripts/creation.sql
@@ -14,7 +14,7 @@ CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
 -- 1) Seguridad: Roles y Permisos
 -- =============================
 CREATE TABLE roles (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   name VARCHAR(50) UNIQUE NOT NULL,
   -- SUPER_ADMIN, PSYCHOLOGIST, ADMIN_STAFF, PATIENT, etc.
   label VARCHAR(100),
@@ -24,7 +24,7 @@ CREATE TABLE roles (
 );
 
 CREATE TABLE permissions (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   code VARCHAR(120) UNIQUE NOT NULL,
   -- VIEW_PATIENT_CLINICAL_RECORDS, GENERATE_INVOICES, etc.
   description TEXT,
@@ -34,8 +34,8 @@ CREATE TABLE permissions (
 
 -- Relación many-to-many Rol <-> Permiso
 CREATE TABLE role_permissions (
-  role_id BIGINT NOT NULL REFERENCES roles(id) ON DELETE CASCADE,
-  permission_id BIGINT NOT NULL REFERENCES permissions(id) ON DELETE CASCADE,
+  role_id INTEGER NOT NULL REFERENCES roles(id) ON DELETE CASCADE,
+  permission_id INTEGER NOT NULL REFERENCES permissions(id) ON DELETE CASCADE,
   created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
   PRIMARY KEY (role_id, permission_id)
 );
@@ -44,11 +44,11 @@ CREATE TABLE role_permissions (
 -- 2) Usuarios / Identidad
 -- =============================
 CREATE TABLE users (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   email VARCHAR(255) UNIQUE NOT NULL,
   username VARCHAR(100) UNIQUE,
   password_hash VARCHAR(255) NOT NULL,
-  role_id BIGINT NOT NULL REFERENCES roles(id),
+  role_id INTEGER NOT NULL REFERENCES roles(id),
   is_active BOOLEAN NOT NULL DEFAULT TRUE,
   two_fa_enabled BOOLEAN NOT NULL DEFAULT FALSE,
   two_fa_secret VARCHAR(255),
@@ -67,7 +67,7 @@ CREATE INDEX idx_users_role_id ON users(role_id);
 -- 3) Catálogos: Áreas y Especialidades
 -- =============================
 CREATE TABLE areas (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   name VARCHAR(120) UNIQUE NOT NULL,
   description TEXT,
   created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
@@ -75,7 +75,7 @@ CREATE TABLE areas (
 );
 
 CREATE TABLE specialties (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   name VARCHAR(120) UNIQUE NOT NULL,
   description TEXT,
   created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
@@ -86,8 +86,8 @@ CREATE TABLE specialties (
 -- 4) Empleados
 -- =============================
 CREATE TABLE employees (
-  id BIGSERIAL PRIMARY KEY,
-  user_id BIGINT UNIQUE REFERENCES users(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  user_id INTEGER UNIQUE REFERENCES users(id) ON DELETE
   SET
     NULL,
     first_name VARCHAR(100) NOT NULL,
@@ -96,7 +96,7 @@ CREATE TABLE employees (
     -- PSYCHOLOGIST, PSYCHIATRIST, TECHNICIAN, MAINTENANCE, ADMIN_STAFF
     license_number VARCHAR(100),
     -- aplica a profesionales
-    area_id BIGINT REFERENCES areas(id) ON DELETE
+    area_id INTEGER REFERENCES areas(id) ON DELETE
   SET
     NULL,
     base_salary NUMERIC(12, 2) NOT NULL DEFAULT 0,
@@ -115,8 +115,8 @@ CREATE INDEX idx_employees_area_id ON employees(area_id);
 CREATE INDEX idx_employees_type ON employees(employee_type);
 
 CREATE TABLE employee_specialties (
-  employee_id BIGINT NOT NULL REFERENCES employees(id) ON DELETE CASCADE,
-  specialty_id BIGINT NOT NULL REFERENCES specialties(id) ON DELETE CASCADE,
+  employee_id INTEGER NOT NULL REFERENCES employees(id) ON DELETE CASCADE,
+  specialty_id INTEGER NOT NULL REFERENCES specialties(id) ON DELETE CASCADE,
   PRIMARY KEY (employee_id, specialty_id)
 );
 
@@ -124,8 +124,8 @@ CREATE TABLE employee_specialties (
 -- 5) Pacientes
 -- =============================
 CREATE TABLE patients (
-  id BIGSERIAL PRIMARY KEY,
-  user_id BIGINT UNIQUE REFERENCES users(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  user_id INTEGER UNIQUE REFERENCES users(id) ON DELETE
   SET
     NULL,
     first_name VARCHAR(100) NOT NULL,
@@ -154,15 +154,15 @@ CREATE INDEX idx_patients_lastname ON patients(last_name);
 -- 6) Historia Clínica (según plantilla)
 -- =============================
 CREATE TABLE clinical_records (
-  id BIGSERIAL PRIMARY KEY,
-  patient_id BIGINT NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  patient_id INTEGER NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
   record_number VARCHAR(50) UNIQUE,
   -- Número de Historia Clínica
   institution_name VARCHAR(150),
   service VARCHAR(120),
   -- Psicología clínica, neuropsicología, etc.
   opening_date DATE,
-  responsible_employee_id BIGINT REFERENCES employees(id) ON DELETE
+  responsible_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     responsible_license VARCHAR(100),
@@ -181,8 +181,8 @@ CREATE INDEX idx_clinical_records_patient ON clinical_records(patient_id);
 
 -- Antecedentes
 CREATE TABLE clinical_backgrounds (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
   family_structure TEXT,
   significant_family_relations TEXT,
   family_psych_history TEXT,
@@ -202,8 +202,8 @@ CREATE TABLE clinical_backgrounds (
 
 -- Evaluación inicial (Likert + observaciones)
 CREATE TABLE initial_assessments (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
   mood_level SMALLINT,
   anxiety_level SMALLINT,
   social_functioning SMALLINT,
@@ -241,8 +241,8 @@ CREATE TABLE initial_assessments (
 
 -- Pruebas psicológicas aplicadas
 CREATE TABLE psychological_tests (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
   test_name VARCHAR(150) NOT NULL,
   application_date DATE,
   result_value VARCHAR(100),
@@ -256,8 +256,8 @@ CREATE INDEX idx_psych_tests_record ON psychological_tests(clinical_record_id);
 
 -- Impresión diagnóstica
 CREATE TABLE diagnostic_impressions (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
   main_diagnosis_code VARCHAR(50),
   main_diagnosis_description TEXT,
   secondary_diagnosis_code VARCHAR(50),
@@ -277,8 +277,8 @@ CREATE TABLE diagnostic_impressions (
 
 -- Plan de intervención
 CREATE TABLE intervention_plans (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
   short_term_goal TEXT,
   medium_term_goal TEXT,
   long_term_goal TEXT,
@@ -297,9 +297,9 @@ CREATE TABLE intervention_plans (
 
 -- Notas por sesión / progreso
 CREATE TABLE sessions (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
-  professional_id BIGINT REFERENCES employees(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
+  professional_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     session_datetime TIMESTAMPTZ NOT NULL,
@@ -323,8 +323,8 @@ CREATE INDEX idx_sessions_professional ON sessions(professional_id);
 
 -- Evaluaciones periódicas
 CREATE TABLE periodic_evaluations (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
   evaluation_date DATE NOT NULL,
   evaluation_type VARCHAR(30),
   -- Parcial, Seguimiento, Final
@@ -347,8 +347,8 @@ CREATE INDEX idx_periodic_eval_record ON periodic_evaluations(clinical_record_id
 
 -- Alta terapéutica
 CREATE TABLE therapeutic_discharges (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL UNIQUE REFERENCES clinical_records(id) ON DELETE CASCADE,
   discharge_date DATE NOT NULL,
   discharge_reason VARCHAR(150),
   discharge_status TEXT,
@@ -363,12 +363,12 @@ CREATE TABLE therapeutic_discharges (
 
 -- Tareas asignadas al paciente
 CREATE TABLE patient_tasks (
-  id BIGSERIAL PRIMARY KEY,
-  patient_id BIGINT NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
-  clinical_record_id BIGINT REFERENCES clinical_records(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  patient_id INTEGER NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
+  clinical_record_id INTEGER REFERENCES clinical_records(id) ON DELETE
   SET
     NULL,
-    assigned_by_employee_id BIGINT REFERENCES employees(id) ON DELETE
+    assigned_by_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     title VARCHAR(200) NOT NULL,
@@ -385,9 +385,9 @@ CREATE INDEX idx_patient_tasks_patient ON patient_tasks(patient_id);
 
 -- Advertencias / alertas
 CREATE TABLE patient_alerts (
-  id BIGSERIAL PRIMARY KEY,
-  patient_id BIGINT NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
-  created_by_employee_id BIGINT REFERENCES employees(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  patient_id INTEGER NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
+  created_by_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     alert_type VARCHAR(50) NOT NULL,
@@ -403,10 +403,10 @@ CREATE INDEX idx_patient_alerts_patient ON patient_alerts(patient_id);
 
 -- Notas confidenciales
 CREATE TABLE confidential_notes (
-  id BIGSERIAL PRIMARY KEY,
-  patient_id BIGINT NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
-  clinical_record_id BIGINT NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
-  author_employee_id BIGINT REFERENCES employees(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  patient_id INTEGER NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
+  clinical_record_id INTEGER NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
+  author_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     content TEXT NOT NULL,
@@ -417,13 +417,13 @@ CREATE INDEX idx_conf_notes_record ON confidential_notes(clinical_record_id);
 
 -- Adjuntos clínicos
 CREATE TABLE clinical_attachments (
-  id BIGSERIAL PRIMARY KEY,
-  clinical_record_id BIGINT NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  clinical_record_id INTEGER NOT NULL REFERENCES clinical_records(id) ON DELETE CASCADE,
   attachment_type VARCHAR(50) NOT NULL,
   -- CONSENT_FORM, TEST_RESULT, COMMUNICATION, etc.
   file_path TEXT NOT NULL,
   description TEXT,
-  uploaded_by_employee_id BIGINT REFERENCES employees(id) ON DELETE
+  uploaded_by_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     uploaded_at TIMESTAMPTZ NOT NULL DEFAULT now()
@@ -435,12 +435,12 @@ CREATE INDEX idx_clinical_attach_record ON clinical_attachments(clinical_record_
 -- 7) Citas
 -- =============================
 CREATE TABLE appointments (
-  id BIGSERIAL PRIMARY KEY,
-  patient_id BIGINT NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
-  professional_id BIGINT REFERENCES employees(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  patient_id INTEGER NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
+  professional_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
-    area_id BIGINT REFERENCES areas(id) ON DELETE
+    area_id INTEGER REFERENCES areas(id) ON DELETE
   SET
     NULL,
     appointment_type VARCHAR(50),
@@ -468,7 +468,7 @@ CREATE INDEX idx_appointments_start ON appointments(start_datetime);
 ALTER TABLE
   sessions
 ADD
-  COLUMN appointment_id BIGINT REFERENCES appointments(id) ON DELETE
+  COLUMN appointment_id INTEGER REFERENCES appointments(id) ON DELETE
 SET
   NULL;
 
@@ -478,16 +478,16 @@ CREATE INDEX idx_sessions_appointment ON sessions(appointment_id);
 -- 8) Facturación / Pagos
 -- =============================
 CREATE TABLE payment_methods (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   name VARCHAR(80) UNIQUE NOT NULL,
   created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
 
 CREATE TABLE services (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   code VARCHAR(50) UNIQUE NOT NULL,
   name VARCHAR(150) NOT NULL,
-  area_id BIGINT REFERENCES areas(id) ON DELETE
+  area_id INTEGER REFERENCES areas(id) ON DELETE
   SET
     NULL,
     default_price NUMERIC(12, 2) NOT NULL DEFAULT 0,
@@ -498,10 +498,10 @@ CREATE TABLE services (
 );
 
 CREATE TABLE invoices (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   invoice_number VARCHAR(60) UNIQUE NOT NULL,
-  patient_id BIGINT NOT NULL REFERENCES patients(id) ON DELETE RESTRICT,
-  created_by_employee_id BIGINT REFERENCES employees(id) ON DELETE
+  patient_id INTEGER NOT NULL REFERENCES patients(id) ON DELETE RESTRICT,
+  created_by_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     invoice_date DATE NOT NULL DEFAULT CURRENT_DATE,
@@ -526,7 +526,7 @@ CREATE INDEX idx_invoices_status ON invoices(status);
 -- 9) Inventario
 -- =============================
 CREATE TABLE products (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   sku VARCHAR(80) UNIQUE,
   name VARCHAR(150) NOT NULL,
   description TEXT,
@@ -551,12 +551,12 @@ CREATE INDEX idx_products_type ON products(product_type);
 
 -- Items de factura (servicios o productos)
 CREATE TABLE invoice_items (
-  id BIGSERIAL PRIMARY KEY,
-  invoice_id BIGINT NOT NULL REFERENCES invoices(id) ON DELETE CASCADE,
-  service_id BIGINT REFERENCES services(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  invoice_id INTEGER NOT NULL REFERENCES invoices(id) ON DELETE CASCADE,
+  service_id INTEGER REFERENCES services(id) ON DELETE
   SET
     NULL,
-    product_id BIGINT REFERENCES products(id) ON DELETE
+    product_id INTEGER REFERENCES products(id) ON DELETE
   SET
     NULL,
     description TEXT,
@@ -581,9 +581,9 @@ CREATE INDEX idx_invoice_items_invoice ON invoice_items(invoice_id);
 
 -- Pagos
 CREATE TABLE payments (
-  id BIGSERIAL PRIMARY KEY,
-  invoice_id BIGINT NOT NULL REFERENCES invoices(id) ON DELETE CASCADE,
-  payment_method_id BIGINT REFERENCES payment_methods(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  invoice_id INTEGER NOT NULL REFERENCES invoices(id) ON DELETE CASCADE,
+  payment_method_id INTEGER REFERENCES payment_methods(id) ON DELETE
   SET
     NULL,
     amount NUMERIC(12, 2) NOT NULL,
@@ -597,18 +597,18 @@ CREATE INDEX idx_payments_invoice ON payments(invoice_id);
 
 -- Movimientos de inventario
 CREATE TABLE inventory_movements (
-  id BIGSERIAL PRIMARY KEY,
-  product_id BIGINT NOT NULL REFERENCES products(id) ON DELETE RESTRICT,
+  id SERIAL PRIMARY KEY,
+  product_id INTEGER NOT NULL REFERENCES products(id) ON DELETE RESTRICT,
   movement_type VARCHAR(20) NOT NULL,
   -- IN, OUT, ADJUSTMENT
   quantity INTEGER NOT NULL,
   movement_date TIMESTAMPTZ NOT NULL DEFAULT now(),
   reason VARCHAR(50),
   -- PURCHASE, SALE, PRESCRIPTION_DELIVERY, ADJUSTMENT, etc.
-  related_invoice_item_id BIGINT REFERENCES invoice_items(id) ON DELETE
+  related_invoice_item_id INTEGER REFERENCES invoice_items(id) ON DELETE
   SET
     NULL,
-    created_by_employee_id BIGINT REFERENCES employees(id) ON DELETE
+    created_by_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
@@ -622,12 +622,12 @@ CREATE INDEX idx_inventory_movements_product ON inventory_movements(product_id);
 -- 10) Prescripciones / Entregas de medicación
 -- =============================
 CREATE TABLE prescriptions (
-  id BIGSERIAL PRIMARY KEY,
-  patient_id BIGINT NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
-  clinical_record_id BIGINT REFERENCES clinical_records(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  patient_id INTEGER NOT NULL REFERENCES patients(id) ON DELETE CASCADE,
+  clinical_record_id INTEGER REFERENCES clinical_records(id) ON DELETE
   SET
     NULL,
-    prescribed_by_employee_id BIGINT REFERENCES employees(id) ON DELETE
+    prescribed_by_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
@@ -637,9 +637,9 @@ CREATE TABLE prescriptions (
 CREATE INDEX idx_prescriptions_patient ON prescriptions(patient_id);
 
 CREATE TABLE prescription_items (
-  id BIGSERIAL PRIMARY KEY,
-  prescription_id BIGINT NOT NULL REFERENCES prescriptions(id) ON DELETE CASCADE,
-  product_id BIGINT NOT NULL REFERENCES products(id) ON DELETE RESTRICT,
+  id SERIAL PRIMARY KEY,
+  prescription_id INTEGER NOT NULL REFERENCES prescriptions(id) ON DELETE CASCADE,
+  product_id INTEGER NOT NULL REFERENCES products(id) ON DELETE RESTRICT,
   dose VARCHAR(80),
   frequency VARCHAR(80),
   duration_days INTEGER,
@@ -649,9 +649,9 @@ CREATE TABLE prescription_items (
 CREATE INDEX idx_prescription_items_presc ON prescription_items(prescription_id);
 
 CREATE TABLE medication_deliveries (
-  id BIGSERIAL PRIMARY KEY,
-  prescription_item_id BIGINT NOT NULL REFERENCES prescription_items(id) ON DELETE CASCADE,
-  delivered_by_employee_id BIGINT REFERENCES employees(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  prescription_item_id INTEGER NOT NULL REFERENCES prescription_items(id) ON DELETE CASCADE,
+  delivered_by_employee_id INTEGER REFERENCES employees(id) ON DELETE
   SET
     NULL,
     quantity INTEGER NOT NULL,
@@ -663,7 +663,7 @@ CREATE TABLE medication_deliveries (
 -- 11) Nómina
 -- =============================
 CREATE TABLE payroll_periods (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   period_start DATE NOT NULL,
   period_end DATE NOT NULL,
   status VARCHAR(20) NOT NULL DEFAULT 'OPEN',
@@ -675,9 +675,9 @@ CREATE TABLE payroll_periods (
 );
 
 CREATE TABLE payroll_records (
-  id BIGSERIAL PRIMARY KEY,
-  employee_id BIGINT NOT NULL REFERENCES employees(id) ON DELETE RESTRICT,
-  period_id BIGINT NOT NULL REFERENCES payroll_periods(id) ON DELETE RESTRICT,
+  id SERIAL PRIMARY KEY,
+  employee_id INTEGER NOT NULL REFERENCES employees(id) ON DELETE RESTRICT,
+  period_id INTEGER NOT NULL REFERENCES payroll_periods(id) ON DELETE RESTRICT,
   base_salary_amount NUMERIC(12, 2) NOT NULL DEFAULT 0,
   sessions_count INTEGER NOT NULL DEFAULT 0,
   sessions_amount NUMERIC(12, 2) NOT NULL DEFAULT 0,
@@ -697,8 +697,8 @@ CREATE INDEX idx_payroll_records_employee ON payroll_records(employee_id);
 -- 12) Auditoría
 -- =============================
 CREATE TABLE audit_logs (
-  id BIGSERIAL PRIMARY KEY,
-  user_id BIGINT REFERENCES users(id) ON DELETE
+  id SERIAL PRIMARY KEY,
+  user_id INTEGER REFERENCES users(id) ON DELETE
   SET
     NULL,
     action VARCHAR(20) NOT NULL,
